% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_bootstrap.R
\name{boot_corrected_auc}
\alias{boot_corrected_auc}
\title{AUC bias correction}
\usage{
boot_corrected_auc(formula, data, B = 200, seed = NULL)
}
\arguments{
\item{formula}{A logistic regression formula used which need to be corrected}

\item{B}{Bootstrap times default is 200}

\item{seed}{If you want to set a seed default is NULL}

\item{df_input}{Data}
}
\value{
bootstrap corrected AUC and CI
}
\description{
This function correct logistic regression AUC by using Efron's bootstrap optimism method
}
\examples{
boot_corrected_auc(
 rec5y_yn ~ kps_lab + pdsids_lab + resid_lab1 + bev_lab + brcaparp_lab + rb1_lab,
 data = logis_mv1_db,
 B = 200,
 seed = 304
)
dxdt = purrr::map_vec(dxdt_orig, date_format)

}
\references{
Efron, B., & Tibshirani, R. J. (1993). An Introduction to the Bootstrap. Chapman & Hall/CRC.

Harrell FE Jr. Regression Modeling Strategies: With Applications to Linear Models, Logistic and Ordinal Regression, and Survival Analysis. 2nd ed. Cham, Switzerland: Springer; 2015.

Harrell Jr FE, Lee KL, Mark DB. Mutlivariable prognostic models: issues in developing models evaluating assumptions and adequacy, and measuring and reducing errors. Statistics in Medicine 1996;15:361–87.
}
