% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{date_format}
\alias{date_format}
\title{Format_Date}
\usage{
date_format(date_input)
}
\arguments{
\item{x}{A date value which is inputted as or as not the format as date}
}
\value{
a formated date
}
\description{
This function takes whatever assuming to be date value_text,number,date transform them into the correct date format
}
\examples{
date_format(34567)
dxdt = purrr::map_vec(dxdt_orig, date_format)

}
