% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_funcs.R
\name{dt_check}
\alias{dt_check}
\title{Check date co-variables' relationship}
\usage{
dt_check(dt.mat, small.col, larger.col, equal.show, direct_out)
}
\arguments{
\item{dt.mat}{A dataframe include all the dates columns in your data}

\item{small.col}{A date column which should show earlier}

\item{larger.col}{A date column which should show larger}

\item{equal.show}{A logic inputted vector as TRUE or FALSE-whether to show the eaqual date}

\item{direct_out}{A logic input vector as TRUE or FALSE-whether to output the error in console}
}
\value{
conflict date error
}
\description{
This function check if one date variable_which should be earlier then the other date variables
}
\examples{
date_check(dt_data, small.col=req(2,3), larger.col=c(4,5,6), equal.show=rep(FALSE,3), direct_out=rep(TRUE,3))

}
