% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_label.R
\name{factor_label}
\alias{factor_label}
\title{pvalue format}
\usage{
factor_label(x, levels, labels, zq_label)
}
\arguments{
\item{x}{variable name}

\item{levels}{variable levels}

\item{labels}{each levels' label}

\item{zq_label}{this variable label}
}
\value{
labeled variable (used in mutation())
}
\description{
This function format pvalue turning it to character
}
\examples{
bmi_lab = factor_label(bmi, leves=c(0,1), labels=c("Not_obese", "Obese"), zq_label="Obesity")

}
