% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_funcs.R
\docType{data}
\name{msk_color}
\alias{msk_color}
\title{Some color schema}
\format{
An object of class \code{list} of length 2.
}
\usage{
msk_color
}
\description{
This create some color schema
}
\keyword{datasets}
