% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_format.R
\name{perm_pv_change}
\alias{perm_pv_change}
\title{Change to permutation pv}
\usage{
perm_pv_change(input.df, input.uv.df, change_vars, change_to_pvs)
}
\arguments{
\item{change_vars}{variables which pvalue should be changed to permutation pv or ''}

\item{change_to_pvs}{permutation pvalue or NA}

\item{infput.df}{Original data with which the univariate survival analyses were done}

\item{infput.uv.df}{The program generated univariate analyses result data_frame}
}
\value{
another survival analyses data frame-replacing original pvalue to be permutation pvalue
}
\description{
This function replace the program generating log-rank with permutation log-rank test pvalue (which should be running separately)
}
\examples{
perm_pv_change(input.df=latrec.wk, input.uv.df=pfs.uv[[1]], change_vars=c('race_lab','hist_lab'), changeto_pvs=c(0.9865, 0.0005))

}
