% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_format.R
\name{pv_format}
\alias{pv_format}
\title{pvalue format}
\usage{
pv_format(pv_input)
}
\arguments{
\item{pv_input}{}
}
\value{
a character pvalue(<0.001) or 3 digits pvalue
}
\description{
This function format pvalue turning it to character
}
\examples{
pv_char = purrr::map_vec(pv_input, pv_format)

}
